extract UI code from TranscriptAnnotations
