+++ reascripts/ReaSpeech/tests/TestReaSpeechUI.lua
@@ -22,6 +22,7 @@ require('source/ReaSpeechWidgets')
 require('source/ReaSpeechWorker')
 require('source/Transcript')
 require('source/TranscriptAnnotations')
+require('source/TranscriptAnnotationsUI')
 require('source/TranscriptEditor')
 require('source/TranscriptExporter')
 require('source/TranscriptUI')
